+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Simply replace the version numbers depending on which set of changes you wish to
   * Bumped version requirement for the PHPUnit dev-dependency to 11.1
   * Updated `tests` github workflow to add uploading coverage data to Codecov.io
   * Added mend.io's Renovate to workflow
+  * Added `Esi\Clock` to dependencies, updated `Esi\Utility\Dates` accordingly.
 
 ### 2.0.0 (2024-03-04)